Fix chisquare test f_expected for latest scipy version.

+++ tests/test_fldr.py
@@ -17,7 +17,7 @@ from fldr import fldr_sample
 
 
 def get_chisquare_pval(p_target, samples):
-    f_expected = [int(len(samples)*p) for p in p_target]
+    f_expected = [len(samples)*p for p in p_target]
     counts = Counter(samples)
     f_actual = [counts[k] for k in range(len(p_target))]
     return chisquare(f_expected, f_actual)[1]
